+++ src/app/components/Results/Results.jsx
@@ -211,8 +211,11 @@ class Results extends React.Component {
 
   render() {
     const results = this.getList(this.state.searchResults);
-    const resultsNumberSuggestion = (results.length === 0) ?
-      'No items were found' : `We found about ${this.props.amount} results.`;
+    let resultsNumberSuggestion = (results.length === 0) ?
+      'No items were found' : `Found about ${this.props.amount.toLocaleString()} results for "${this.props.searchKeyword}"`;
+    if (this.props.selectedFacet !== undefined && this.props.selectedFacet !== ''){
+      resultsNumberSuggestion += ` in ${this.props.selectedFacet.replace('_', ' ')}`;
+    }
     const resultMessageClass = (results.length === 0) ?
       'noResultMessage' : `${this.props.className}-length`;
 
@@ -220,8 +223,6 @@ class Results extends React.Component {
       <div className={`${this.props.className}-wrapper`}>
         <p
           className={resultMessageClass}
-          role="alert"
-          aria-atomic="true"
           aria-live="polite"
         >
           {resultsNumberSuggestion}